requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
cryptography>=42.0.0
idna==3.11
numpy==2.3.5
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
scipy==1.16.3
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
streamlit>=1.30.0
altair>=5.0.0
python-dotenv>=1.0.0
